requirements.txt
langchain
langchain-core
langchain-community
python-dotenv
transformers
pandas
FastAPI
uvicorn
datasets
langchain-openai
nltk
langchain-groq
metaphone
pinecone-client
rapidfuzz
langchain-ollama
pinecone-text